+++ src/extension.ts
@@ -27,6 +27,7 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand('voxpilot.pushToTalk', () => engine?.quickCapture()),
     vscode.commands.registerCommand('voxpilot.selectModel', () => engine?.selectModel()),
     vscode.commands.registerCommand('voxpilot.selectAudioDevice', () => engine?.selectAudioDevice()),
+    vscode.commands.registerCommand('voxpilot.transcriptHistory', () => engine?.showTranscriptHistory()),
     vscode.commands.registerCommand('voxpilot.sendToChat', () => engine?.sendLastToChat()),
     vscode.commands.registerCommand('voxpilot.clearCache', () => clearCache(context)),
     statusBar,